Extract helper class for moving nodes arround
